Safari-609.1.11.1/LayoutTests/http/tests/performance/performance-resource-timing-xhr-single-entry-expected.txt

PASS foundResource is 1
PASS initiator is "xmlhttprequest"
